Add progress log for scrape_pagelist and scrape_all_pages

+++ trope-download/src/cmd/download_pagelist.rs
@@ -16,8 +16,8 @@ pub fn save_pagelist(args: trope_lib::TropeDownloadPagelist) -> Result<(), Box<d
   let page_dir = trope_lib::dl_page_dir(&ns);
 
   // Inclusive
-  let beg_record = 0.max(args.beg_record);
-  let end_record = 0.max(args.end_record);
+  let beg_record = 0.max(args.beg_record) as usize;
+  let end_record = 0.max(args.end_record) as usize;
   if end_record < beg_record {
     panic!("end_record should not be less than beg_record");
   }
@@ -40,11 +40,11 @@ pub fn save_pagelist(args: trope_lib::TropeDownloadPagelist) -> Result<(), Box<d
   log::info!("Downloading {} to {} of {} records...", beg_record, end_record, tot_records);
 
   // Page request loop with peekable iterator
-  let mut tup_iter = (beg_record..end_record+1).zip(record_iter.skip(beg_record as usize)).peekable();
+  let mut tup_iter = (beg_record..end_record+1).zip(record_iter.skip(beg_record)).peekable();
   let mut progress_tick = 0;
   while let Some((idx, record)) = tup_iter.next() {
 
-    if args.progress && (100 * idx > (progress_tick * tot_records) as u64) {
+    if args.progress && (100 * idx > progress_tick * tot_records) {
       log::warn!("Progress: {}%", progress_tick);
       progress_tick += 1;
     }
